requirements.txt
Django==4.1.1
djangorestframework
django-cors-headers
pandas
numpy
seaborn
matplotlib
scikit-learn
nltk
gensim
tensorflow-cpu==2.16.1
Pillow
pytesseract
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
